Move Continuity compaction into plugin hooks
