+++ code_puppy/agents/_compaction.py
@@ -32,6 +32,7 @@ from code_puppy.agents._history import (
     prune_interrupted_tool_calls,
 )
 from code_puppy.callbacks import (
+    on_compact_message_history,
     on_message_history_processor_end,
     on_message_history_processor_start,
 )
@@ -316,30 +317,28 @@ def compact(
     update_spinner_context(context_summary)
 
     strategy = get_compaction_strategy()
-    if strategy == "continuity":
-        # This cannot currently live as a regular Code Puppy plugin without a
-        # new core extension point: compaction owns history-processor mutation
-        # and must preserve pydantic-ai tool-call/tool-return ordering.
-        from code_puppy.agents.continuity_compaction import compact_continuity
-
-        result_messages, summarized_messages = compact_continuity(
-            agent=agent,
-            messages=messages,
-            model_max=model_max,
-            context_overhead=context_overhead,
-            model_name=model_name,
-            force=force,
-        )
-        final_token_count = sum(
-            estimate_tokens_for_message(m, model_name) for m in result_messages
-        )
-        final_summary = SpinnerBase.format_context_info(
-            final_token_count,
-            model_max,
-            final_token_count / model_max if model_max else 0.0,
-        )
-        update_spinner_context(final_summary)
+    plugin_result = _run_plugin_compaction(
+        strategy=strategy,
+        agent=agent,
+        messages=messages,
+        model_max=model_max,
+        context_overhead=context_overhead,
+        model_name=model_name,
+        force=force,
+        total_tokens=total_tokens,
+        proportion_used=proportion_used,
+    )
+    if plugin_result is not None:
+        result_messages, summarized_messages = plugin_result
+        _update_final_spinner(result_messages, model_name, model_max)
         return result_messages, summarized_messages
+
+    if strategy not in {"summarization", "truncation"}:
+        emit_warning(
+            f"Compaction strategy '{strategy}' was not handled by any plugin; "
+            "falling back to truncation for this compaction cycle."
+        )
+        strategy = "truncation"
 
     threshold = get_compaction_threshold()
     if not force and proportion_used <= threshold:
@@ -387,8 +386,58 @@ def compact(
                 filtered, protected_tokens, model_name
             )
 
+    _update_final_spinner(result_messages, model_name, model_max)
+
+    return result_messages, summarized_messages
+
+
+def _run_plugin_compaction(
+    *,
+    strategy: str,
+    agent: Any,
+    messages: List[ModelMessage],
+    model_max: int,
+    context_overhead: int,
+    model_name: Optional[str],
+    force: bool,
+    total_tokens: int,
+    proportion_used: float,
+) -> Tuple[List[ModelMessage], List[ModelMessage]] | None:
+    """Return plugin compaction output when a plugin handles the strategy."""
+    results = on_compact_message_history(
+        strategy=strategy,
+        agent=agent,
+        messages=messages,
+        model_max=model_max,
+        context_overhead=context_overhead,
+        model_name=model_name,
+        force=force,
+        total_tokens=total_tokens,
+        proportion_used=proportion_used,
+    )
+    for result in results:
+        if not isinstance(result, dict) or not result.get("handled"):
+            continue
+        result_messages = result.get("messages")
+        if not isinstance(result_messages, list):
+            emit_warning(
+                f"Compaction plugin for '{strategy}' returned no message list; ignoring."
+            )
+            continue
+        dropped = result.get("dropped_messages", result.get("dropped", []))
+        if not isinstance(dropped, list):
+            dropped = []
+        return result_messages, dropped
+    return None
+
+
+def _update_final_spinner(
+    messages: List[ModelMessage],
+    model_name: Optional[str],
+    model_max: int,
+) -> None:
     final_token_count = sum(
-        estimate_tokens_for_message(m, model_name) for m in result_messages
+        estimate_tokens_for_message(m, model_name) for m in messages
     )
     final_summary = SpinnerBase.format_context_info(
         final_token_count,
@@ -396,8 +445,6 @@ def compact(
         final_token_count / model_max if model_max else 0.0,
     )
     update_spinner_context(final_summary)
-
-    return result_messages, summarized_messages
 
 
 def _strip_empty_thinking_parts(
