+++ code_puppy/command_line/session_commands.py
@@ -74,7 +74,7 @@ def handle_session_command(command: str) -> bool:
 def handle_compact_command(command: str) -> bool:
     """Compact message history using configured strategy."""
     from code_puppy.agents.agent_manager import get_current_agent
-    from code_puppy.config import get_compaction_strategy, get_protected_token_count
+    from code_puppy.config import get_compaction_strategy
     from code_puppy.messaging import emit_error, emit_info, emit_success, emit_warning
 
     try:
@@ -89,32 +89,20 @@ def handle_compact_command(command: str) -> bool:
             current_agent.estimate_tokens_for_message(m) for m in history
         )
         compaction_strategy = get_compaction_strategy()
-        protected_tokens = get_protected_token_count()
         emit_info(
             f"🤔 Compacting {len(history)} messages using {compaction_strategy} strategy... (~{before_tokens} tokens)"
         )
 
         current_agent = get_current_agent()
-        if compaction_strategy == "truncation":
-            from code_puppy.agents._compaction import truncate
+        from code_puppy.agents._compaction import compact
 
-            compacted = truncate(history, protected_tokens)
-            summarized_messages = []  # No summarization in truncation mode
-        elif compaction_strategy == "continuity":
-            from code_puppy.agents._compaction import compact
-
-            compacted, summarized_messages = compact(
-                current_agent,
-                history,
-                current_agent._get_model_context_length(),
-                current_agent._estimate_context_overhead(),
-                force=True,
-            )
-        else:
-            # Default to summarization
-            compacted, summarized_messages = current_agent.summarize_messages(
-                history, with_protection=True
-            )
+        compacted, summarized_messages = compact(
+            current_agent,
+            history,
+            current_agent._get_model_context_length(),
+            current_agent._estimate_context_overhead(),
+            force=True,
+        )
 
         if not compacted:
             emit_error("Compaction failed. History unchanged.")
@@ -149,166 +137,6 @@ def handle_compact_command(command: str) -> bool:
     except Exception as e:
         emit_error(f"/compact error: {e}")
         return True
-
-
-@register_command(
-    name="continuity",
-    description="Show continuity compaction memory and archives",
-    usage="/continuity [show|tasks|diagnostics|archives search <query>|archives show <id>]",
-    category="session",
-)
-def handle_continuity_command(command: str) -> bool:
-    """Inspect continuity memory state for the current session."""
-    from code_puppy.agents.agent_manager import get_current_agent
-    from code_puppy.agents.continuity_compaction.storage import (
-        archive_preview,
-        build_archive_index,
-        read_durable_state,
-        read_observation_archive,
-        search_archive_index,
-    )
-    from code_puppy.config import (
-        get_continuity_compaction_archive_retention_count,
-        get_continuity_compaction_archive_retention_days,
-        get_continuity_compaction_archive_retrieval_count,
-        get_continuity_compaction_archive_retrieval_enabled,
-        get_continuity_compaction_predictive_trigger_min_ratio,
-        get_continuity_compaction_semantic_task_detection,
-        get_continuity_compaction_semantic_timeout_seconds,
-    )
-    from code_puppy.messaging import emit_error, emit_info, emit_warning
-
-    tokens = command.split()
-    action = tokens[1].lower() if len(tokens) > 1 else "show"
-
-    try:
-        agent = get_current_agent()
-        state = read_durable_state(agent)
-        archive_index = build_archive_index(agent)
-    except Exception as exc:
-        emit_error(f"/continuity error: {exc}")
-        return True
-
-    if action in {"show", "status"}:
-        if state is None:
-            emit_warning("No continuity memory has been written for this session yet.")
-            return True
-        current_constraints = []
-        for task in state.tasks:
-            if task.task_id == state.current_task_id:
-                current_constraints = task.constraints
-                break
-        lines = [
-            "[bold magenta]Continuity Memory[/bold magenta]",
-            f"Current task: {state.current_task or 'unknown'}",
-            f"Latest request: {state.latest_user_request or 'unknown'}",
-            f"Semantic status: {state.semantic_status or 'unknown'}",
-            f"Archive count: {len(archive_index)}",
-            "Active constraints:",
-        ]
-        constraints = [*state.global_constraints, *current_constraints]
-        lines.extend(f"- {item}" for item in constraints[:12] or ["none"])
-        lines.extend(
-            [
-                "Task ledger:",
-                *_continuity_task_lines(state.tasks, limit=8),
-            ]
-        )
-        emit_info("\n".join(lines))
-        return True
-
-    if action == "tasks":
-        if state is None:
-            emit_warning("No continuity task memory has been written yet.")
-            return True
-        lines = [
-            "[bold magenta]Continuity Tasks[/bold magenta]",
-            *_continuity_task_lines(state.tasks, limit=100),
-        ]
-        emit_info("\n".join(lines))
-        return True
-
-    if action == "diagnostics":
-        lines = [
-            "[bold magenta]Continuity Diagnostics[/bold magenta]",
-            f"semantic_enabled: {get_continuity_compaction_semantic_task_detection()}",
-            f"semantic_timeout_seconds: {get_continuity_compaction_semantic_timeout_seconds()}",
-            f"predictive_trigger_min_ratio: {get_continuity_compaction_predictive_trigger_min_ratio():.3f}",
-            f"archive_retrieval_enabled: {get_continuity_compaction_archive_retrieval_enabled()}",
-            f"archive_retrieval_count: {get_continuity_compaction_archive_retrieval_count()}",
-            f"archive_retention_days: {get_continuity_compaction_archive_retention_days()}",
-            f"archive_retention_count: {get_continuity_compaction_archive_retention_count()}",
-            f"archive_count: {len(archive_index)}",
-        ]
-        if state is not None:
-            lines.extend(
-                [
-                    f"schema_version: {state.schema_version}",
-                    f"last_semantic_status: {state.semantic_status or 'unknown'}",
-                    f"fallback_reason: {state.semantic_error or 'none'}",
-                    f"retrieved_archives: {len(state.retrieved_archive_signals)}",
-                ]
-            )
-        emit_info("\n".join(lines))
-        return True
-
-    if action == "archives":
-        if len(tokens) < 3:
-            emit_warning(
-                "Usage: /continuity archives search <query> or /continuity archives show <id>"
-            )
-            return True
-        archive_action = tokens[2].lower()
-        if archive_action == "search":
-            query = command.split("search", 1)[1].strip() if "search" in command else ""
-            if not query:
-                emit_warning("Usage: /continuity archives search <query>")
-                return True
-            results = search_archive_index(agent, query, limit=10)
-            if not results:
-                emit_info(f"No archive signals matched: {query}")
-                return True
-            lines = [f"[bold magenta]Archive Search[/bold magenta]: {query}"]
-            for item in results:
-                signals = "; ".join((item.get("key_signals") or [])[:2])
-                lines.append(
-                    f"- {item.get('observation_id')} [{item.get('status')}] "
-                    f"{item.get('tool_name')}: {signals or item.get('key_signal') or 'no signal'}"
-                )
-            emit_info("\n".join(lines))
-            return True
-        if archive_action == "show":
-            if len(tokens) < 4:
-                emit_warning("Usage: /continuity archives show <id>")
-                return True
-            record = read_observation_archive(agent, tokens[3])
-            if record is None:
-                emit_warning(f"Archive observation not found: {tokens[3]}")
-                return True
-            emit_info(
-                "[bold magenta]Archive Observation[/bold magenta]\n"
-                + archive_preview(record)
-            )
-            return True
-
-    emit_warning(
-        "Usage: /continuity [show|tasks|diagnostics|archives search <query>|archives show <id>]"
-    )
-    return True
-
-
-def _continuity_task_lines(tasks, *, limit: int) -> list[str]:
-    if not tasks:
-        return ["- none"]
-    lines: list[str] = []
-    for task in tasks[-limit:]:
-        files = ", ".join(task.active_files[:3])
-        files_suffix = f" | files: {files}" if files else ""
-        summary_suffix = f" | {task.summary}" if task.summary else ""
-        lines.append(
-            f"- [{task.status}] {task.task_id}: {task.title}{summary_suffix}{files_suffix}"
-        )
-    return lines
 
 
 @register_command(
