+++ code_puppy/plugins/continuity_compaction/__init__.py
@@ -1,0 +1,5 @@
+"""Continuity-oriented message-history compaction plugin."""
+
+from code_puppy.plugins.continuity_compaction.engine import compact_continuity
+
+__all__ = ["compact_continuity"]
